Remove the processor option for obfuscation and old RegistryGenerator implementation
