+++ toothpick-compiler/src/main/java/toothpick/compiler/common/ToothpickProcessor.java
@@ -29,7 +29,6 @@ import javax.tools.JavaFileObject;
 import toothpick.compiler.common.generators.CodeGenerator;
 import toothpick.compiler.common.generators.targets.ParamInjectionTarget;
 import toothpick.compiler.memberinjector.targets.FieldInjectionTarget;
-import toothpick.compiler.registry.generators.ObfuscationFriendlyRegistryGenerator;
 import toothpick.compiler.registry.generators.RegistryGenerator;
 import toothpick.compiler.registry.targets.RegistryInjectionTarget;
 
@@ -55,11 +54,6 @@ public abstract class ToothpickProcessor extends AbstractProcessor {
    * If this parameter is not passed, no registry is generated.
    */
   public static final String PARAMETER_REGISTRY_PACKAGE_NAME = "toothpick_registry_package_name";
-
-  /**
-   * Parameter to be set to true to ensure Toothpick works with obfuscation.
-   */
-  public static final String PARAMETER_SUPPORT_OBFUSCATION = "toothpick_obfuscation_support";
 
   /**
    * The name of the annotation processor option to exclude classes from the creation of member scopes & factories.
@@ -113,7 +107,6 @@ public abstract class ToothpickProcessor extends AbstractProcessor {
   protected String toothpickExcludeFilters = "java.*,android.*";
   protected Boolean toothpickCrashWhenMethodIsNotPackageVisible;
   protected Set<String> supportedAnnotationTypes = new HashSet<>();
-  protected boolean supportObfuscation;
   private boolean hasAlreadyRun;
 
   @Override
@@ -177,19 +170,6 @@ public abstract class ToothpickProcessor extends AbstractProcessor {
     readOptionRegistryPackageName();
     readOptionRegistryChildrenPackageNames();
     readOptionExcludes();
-    readOptionSupportObfuscation();
-  }
-
-  private void readOptionSupportObfuscation() {
-    Map<String, String> options = processingEnv.getOptions();
-    String supportsObfuscationOption = options.get(PARAMETER_SUPPORT_OBFUSCATION);
-    if (supportsObfuscationOption != null) {
-      supportObfuscation = "true".equals(supportsObfuscationOption);
-    }
-    if (supportObfuscation && !options.containsKey(PARAMETER_REGISTRY_PACKAGE_NAME)) {
-      warning("Option -A%s set to true, but option -A%s is not set. Obfuscation can only work in configuration with registries",
-              PARAMETER_SUPPORT_OBFUSCATION, PARAMETER_REGISTRY_PACKAGE_NAME);
-    }
   }
 
   private void readOptionRegistryPackageName() {
@@ -615,8 +595,6 @@ public abstract class ToothpickProcessor extends AbstractProcessor {
   }
 
   protected CodeGenerator createRegistryGenerator(RegistryInjectionTarget target) {
-    return supportObfuscation
-            ? new ObfuscationFriendlyRegistryGenerator(target, typeUtils)
-            : new RegistryGenerator(target, typeUtils);
+    return new RegistryGenerator(target, typeUtils);
   }
 }
