+++ toothpick-compiler/src/test/java/toothpick/compiler/factory/ProcessorTestUtilities.java
@@ -25,30 +25,15 @@ final class ProcessorTestUtilities {
   }
 
   static Iterable<? extends Processor> factoryProcessors(String toothpickRegistryPackageName, List<String> toothpickRegistryChildrenPackageNameList) {
-    return factoryProcessors(toothpickRegistryPackageName, toothpickRegistryChildrenPackageNameList, "java.*,android.*", false);
+    return factoryProcessors(toothpickRegistryPackageName, toothpickRegistryChildrenPackageNameList, "java.*,android.*");
   }
 
   static Iterable<? extends Processor> factoryProcessors(String toothpickRegistryPackageName,
-                                                         List<String> toothpickRegistryChildrenPackageNameList,
-                                                         boolean supportObfuscation) {
-    return factoryProcessors(toothpickRegistryPackageName, toothpickRegistryChildrenPackageNameList, "java.*,android.*", supportObfuscation);
-  }
-
-  static Iterable<? extends Processor> factoryProcessors(String toothpickRegistryPackageName,
-                                                         List<String> toothpickRegistryChildrenPackageNameList,
-                                                         String toothpickExcludeFilters) {
-    return factoryProcessors(toothpickRegistryPackageName, toothpickRegistryChildrenPackageNameList, toothpickExcludeFilters, false);
-  }
-
-  static Iterable<? extends Processor> factoryProcessors(String toothpickRegistryPackageName,
-                                                         List<String> toothpickRegistryChildrenPackageNameList,
-                                                         String toothpickExcludeFilters,
-                                                         boolean supportObfuscation) {
+                                                         List<String> toothpickRegistryChildrenPackageNameList, String toothpickExcludeFilters) {
     FactoryProcessor factoryProcessor = new FactoryProcessor();
     factoryProcessor.setToothpickRegistryPackageName(toothpickRegistryPackageName);
     factoryProcessor.setToothpickRegistryChildrenPackageNameList(toothpickRegistryChildrenPackageNameList);
     factoryProcessor.setToothpickExcludeFilters(toothpickExcludeFilters);
-    factoryProcessor.setSupportObfuscation(supportObfuscation);
     return Arrays.asList(factoryProcessor);
   }
