--- toothpick-compiler/src/test/java/toothpick/compiler/registry/generators/RegistryGeneratorTestUtilities.java
@@ -1,14 +1,0 @@
-package toothpick.compiler.registry.generators;
-
-public final class RegistryGeneratorTestUtilities {
-  private RegistryGeneratorTestUtilities() {
-  }
-
-  public static void setInjectionTarjetsPerGetterMethod(int injectionTarjetsPerMethod) {
-    RegistryGenerator.injectionTargetsPerGetterMethod = injectionTarjetsPerMethod;
-  }
-
-  public static void setGroupSizeForObfuscationFriendlyGenerator(int groupSize) {
-    ObfuscationFriendlyRegistryGenerator.groupSize = groupSize;
-  }
-}
